requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==9.0.1
PyJWT==2.1.0
python-dotenv==0.19.2
pytz==2020.1
sqlparse==0.3.1
